+++ sincmathsmp/src/androidMain/kotlin/io/github/gallvp/sincmaths/ASincMatrixSolve.kt
@@ -1,0 +1,3 @@
+package io.github.gallvp.sincmaths
+
+actual fun SincMatrix.solve(b: SincMatrix): SincMatrix = this.asSimpleMatrix().solve(b.asSimpleMatrix()).asSincMatrix()