+++ sincmathsmp/src/androidMain/kotlin/io/github/gallvp/sincmaths/ATypeAsSincMatrix.kt
@@ -1,0 +1,5 @@
+package io.github.gallvp.sincmaths
+
+import org.ejml.simple.SimpleMatrix
+
+internal fun SimpleMatrix.asSincMatrix() = SincMatrix(this.ddrm.data, this.numRows, this.numCols)