+++ sincmathsmp/src/commonMain/kotlin/com/github/gallvp/sincmaths/SincMatrixSolve.kt
@@ -1,3 +1,0 @@
-package com.github.gallvp.sincmaths
-
-expect fun SincMatrix.solve(b: SincMatrix): SincMatrix