+++ sincmathsmp/src/commonMain/kotlin/com/github/gallvp/sincmaths/SincMatrixTrigon.kt
@@ -1,5 +1,0 @@
-package com.github.gallvp.sincmaths
-
-expect fun SincMatrix.sin(): SincMatrix
-
-expect fun SincMatrix.cos(): SincMatrix