+++ sincmathsmp/src/commonMain/kotlin/io/github/gallvp/sincmaths/SincMatrixSolve.kt
@@ -1,0 +1,3 @@
+package io.github.gallvp.sincmaths
+
+expect fun SincMatrix.solve(b: SincMatrix): SincMatrix