+++ sincmathsmp/src/cpp/tinyexpr/tinyexpr-jni.cpp
@@ -3,8 +3,8 @@
 
 extern "C"
 JNIEXPORT jdouble JNICALL
-Java_com_github_gallvp_sincmaths_ASincMatrixWorkersKt_parseToDouble(JNIEnv *env, jclass clazz,
-                                                                     jstring expr) {
+Java_io_github_gallvp_sincmaths_ASincMatrixWorkersKt_parseToDouble(JNIEnv *env, jclass clazz,
+                                                                   jstring expr) {
     const char *nativeString = env->GetStringUTFChars(expr, 0);
     double ans = te_interp(nativeString, 0);
     return ans;