+++ sincmathsmp/src/iosMain/kotlin/io/github/gallvp/sincmaths/SincMatrixAsType.kt
@@ -1,0 +1,12 @@
+package io.github.gallvp.sincmaths
+
+actual fun SincMatrix.asArray(): DoubleArray {
+    if (this.isEmpty()) {
+        return doubleArrayOf()
+    }
+
+    require(this.isVector) { "Matrix is not a vector and conversion is invalid" }
+    return this.matrixData
+}
+
+actual fun SincMatrix.asRowMajorArray(): DoubleArray = this.matrixData