+++ sincmathsmp/src/iosMain/kotlin/io/github/gallvp/sincmaths/SincMatrixSolve.kt
@@ -1,0 +1,9 @@
+package io.github.gallvp.sincmaths
+
+actual fun SincMatrix.solve(b: SincMatrix): SincMatrix =
+    solveNonsymSquareSystem(
+        this.matrixData,
+        this.numRows,
+        this.numCols,
+        b.matrixData,
+    ).asSincMatrix(b.numRows, b.numCols)