+++ sincmathsmp/src/iosMain/kotlin/io/github/gallvp/sincmaths/SincMatrixTrigon.kt
@@ -1,0 +1,5 @@
+package io.github.gallvp.sincmaths
+
+actual fun SincMatrix.sin(): SincMatrix = sinOfElementsOfVector(this.matrixData).asSincMatrix(this.numRows, this.numCols)
+
+actual fun SincMatrix.cos(): SincMatrix = cosOfElementsOfVector(this.matrixData).asSincMatrix(this.numRows, this.numCols)