+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Updated ejml-simple to 0.43.1
 + Updated AGP to 8.2.0
 + Added tinyexpr and wavelib libraries as submodules
++ Changed package name to `io.github.gallvp.sincmathsmp`
 + `isVector` now replaced by `isVector`
 + `numRows()` and `numCols()` are now `numRows` and `numCols`
 + `length()` is now `length`